+++ app/models/product.rb
@@ -8,7 +8,7 @@ class Product < ApplicationRecord
     validates :slug, presence: true, uniqueness: true
 
     def set_slug
-        self.slug ||= name.parameterize if name.present?
+        self.slug = name.parameterize if slug.blank? && name.present?
     end
 
     def image_url